refactor(web): group components by feature and consolidate sidebars
Reorganize web/next/src/components by feature/domain so the tree reads the way
the app is built, and collapse cohesive multi-file units into one file.

- common/: access, mode-toggle, devtools, navbar, route-error, route-loading
- shell/: the shared app-shell chrome (was mislabeled dashboard/ and scattered
  across sidebar/)
- dashboard|console|docs/: one sidebar.tsx per area (index barrels dropped)
- merges: docs sidebar 4->1, dashboard sidebar 3->1, shell content 2->1
- keep the full shadcn ui/ kit (shadcn add -a ships the whole registry on
  purpose); the 41 unused primitives are intentional surface, not dead code

Drop redundant Sidebar* prefixes now that folders convey area, and rename the
shared DashboardShell/DashboardHeader to PageShell/PageHeader (console uses them
too). shell/ keeps its Sidebar* names since the sidebar is the subject there.

Update the CLI fork-converter (convert.ts) and its test for the navbar move,
else the /hire strip silently no-ops and forks ship a dead link. Sync the
dashboard/theming/authentication/project-structure docs. Audit report in
.github/audit/2026-07-06-frontend-segregation.md.

No behavior change: tsc, oxlint, oxfmt, and the CLI convert tests all pass.
