+++ web/next/content/docs/getting-started/project-structure.mdx
@@ -39,7 +39,7 @@ Standard App Router, grouped by concern: `(content)/` for docs and blog, `(prote
 
 <Callout type="info" title="Marketing routes">
 
-The `(marketing)/` route group under `src/app/` holds the marketing surfaces: the landing page (`/`) plus the author's own `/hire` and `/resume`. They stay in the canonical repo but `zerostarter init` strips the whole group and writes a fresh `app/page.tsx` in its place. A direct clone that skips init can delete `web/next/src/app/(marketing)`, add its own `app/page.tsx`, and remove the "Hire" link in `web/next/src/components/navbar/home.tsx`.
+The `(marketing)/` route group under `src/app/` holds the marketing surfaces: the landing page (`/`) plus the author's own `/hire` and `/resume`. They stay in the canonical repo but `zerostarter init` strips the whole group and writes a fresh `app/page.tsx` in its place. A direct clone that skips init can delete `web/next/src/app/(marketing)`, add its own `app/page.tsx`, and remove the "Hire" link in `web/next/src/components/common/navbar.tsx`.
 
 </Callout>
