+++ web/next/content/docs/manage/theming.mdx
@@ -20,7 +20,7 @@ You restyle ZeroStarter by editing tokens, not components. Colors, fonts, and ra
 >
 ```
 
-The `ModeToggle` (`web/next/src/components/mode-toggle.tsx`) is a smart three-state toggle: from `system` it jumps to the opposite of the OS preference, flips between `light` and `dark` while they diverge, and snaps back to `system` once the explicit choice matches the OS again.
+The `ModeToggle` (`web/next/src/components/common/mode-toggle.tsx`) is a smart three-state toggle: from `system` it jumps to the opposite of the OS preference, flips between `light` and `dark` while they diverge, and snaps back to `system` once the explicit choice matches the OS again.
 
 ## Design tokens
