+++ web/next/src/app/(protected)/dashboard/page.tsx
@@ -1,13 +1,13 @@
-import { DashboardHeader } from "@/components/dashboard/header"
-import { DashboardShell } from "@/components/dashboard/shell"
+import { PageHeader } from "@/components/shell/content"
+import { PageShell } from "@/components/shell/content"
 
 export default function Page() {
   return (
-    <DashboardShell>
-      <DashboardHeader
+    <PageShell>
+      <PageHeader
         title="Dashboard"
         description="Intentionally empty. Auth, orgs, and the API are wired; this page is where your product begins."
       />
-    </DashboardShell>
+    </PageShell>
   )
 }
