+++ web/next/src/components/common/navbar.tsx
@@ -12,8 +12,8 @@ import Link from "next/link"
 import { usePathname } from "next/navigation"
 import { useEffect, useState } from "react"
 
-import { Access } from "@/components/access"
-import { ModeToggle } from "@/components/mode-toggle"
+import { Access } from "@/components/common/access"
+import { ModeToggle } from "@/components/common/mode-toggle"
 import { Button } from "@/components/ui/button"
 import { Sheet, SheetContent, SheetHeader, SheetTitle, SheetTrigger } from "@/components/ui/sheet"
 import { Spinner } from "@/components/ui/spinner"
