+++ web/next/src/components/shell/adaptive-sidebar.tsx
