--- web/next/src/components/sidebar/dashboard/index.tsx
@@ -1,2 +1,0 @@
-export { SidebarDashboardOrgSwitcher } from "@/components/sidebar/dashboard/org-switcher"
-export { SidebarDashboardUserActions } from "@/components/sidebar/dashboard/user-actions"
