--- web/next/src/components/sidebar/docs/index.tsx
@@ -1,3 +1,0 @@
-export { SidebarDocsContent } from "@/components/sidebar/docs/content"
-export { SidebarDocsFooter } from "@/components/sidebar/docs/footer"
-export { SidebarDocsSearch } from "@/components/sidebar/docs/search"
